Convert `auth` service to an ES class
Refactor this class to new conventions as part of
https://github.com/hypothesis/client/issues/3298.

 - Convert the `auth` service to an ES class
 - Replace the internal `emitter` variable with inheritance
 - Rename the module from `oauth-auth.js` to just `auth.js`. The
   previous name is a vestige from when authorization was migrated from
   cookie-based to OAuth-based
